Break TravelTypeDropdown into component
